Fix bug that caused transition on epsilon
The method Item.getExpectedSymbol() now returns null if the expected
symbol is EPSILON.

Epsilon represents the empty string. In many ways this symbol behaves
like a non-terminal, but the lexer will never return a token
representing 'epsilon'. For this reason should we exclude transitions by
the epsilon token.

Also replaced an old example of a canonical collection by a generated
one.