Add method isVanishable() to Symbol
In order to save memory and speed up the algorithm, items in states will
be combined. The algorithm that will do this will need to know whether
a symbol is vanishable.

This change only adds the isVanishable() method. Terminals, except
epsilon, can never vanish. In subsequent changes I will add an algorithm
that will determine whether this is the case for non-terminals.